requirement.txt
flask
flask-sqlalchemy
flask-session
python-dotenv
flask_bcrypt
marshmallow
apispec
flask_apispec
apispec-webframeworks
flask_restful
flask-cors
flask-mysql
Flask-Migrate
mysqlclient
flask_httpauth
routeros_api
pyjwt
requests
Flask-APScheduler
gunicorn
